Refactor test packages
