+++ src/test/java/me/karakelley/http/http/RequestTest.java
@@ -1,7 +1,5 @@
-package me.karakelley.http;
+package me.karakelley.http.http;
 
-import me.karakelley.http.http.HttpMethod;
-import me.karakelley.http.http.Request;
 import org.junit.jupiter.api.Test;
 
 import java.util.HashMap;
